# server/src/openapi.yaml
openapi: 3.0.3
info:
  title: Practice CRM API
  version: 1.0.0
  description: >
    A real REST API for the Practice CRM training app, built for API-automation
    practice (Postman, RestAssured, pytest, curl, etc.). Records created here
    show up in the CRM UI on next navigation/reload of the relevant page — this
    API is the CRM's single source of truth.


    **Reset**: `POST /api/reset` restores the deterministic seed state at any
    time (no auth required — see that endpoint's description). Seed login:
    `admin@crm.com` / `Pass@123` (also `rep@crm.com`, `viewer@crm.com`,
    `sam@crm.com`, `priya@crm.com` — all deactivated except `priya@crm.com`,
    same password). **Passwords are stored in plaintext by design** — this is
    a training tool with no real user data, and the seed credentials are
    published intentionally for testers to use.


    **Free-tier note**: this server may be deployed on a free host that sleeps
    after inactivity; a cold start re-seeds automatically, so state resets
    after idle periods. This is expected, not a bug.
servers:
  - url: /api
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Request conflicts with the resource's current state (illegal transition, confirmation required, etc.).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
paths:
...
  /tickets/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get: { tags: [Tickets], responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } } }
    put:
      tags: [Tickets]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Updated }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Tickets]
      summary: Delete a ticket. Only Closed tickets may be deleted.
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Ticket is not Closed. }
  /tickets/{id}/transition:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [Tickets]
      summary: Move a ticket's status, validated against the allowed transition map.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [status], properties: { status: { type: string } } }
      responses:
        '200': { description: Transitioned. }
        '409': { $ref: '#/components/responses/Conflict' }
  /tickets/{id}/priority:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [Tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [priority], properties: { priority: { type: string } } }
      responses: { '200': { description: Updated. } }
  /tickets/{id}/comments:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [Tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [text], properties: { text: { type: string } } }
      responses: { '201': { description: Comment added. } }
...